backend/requirement.txt
Django==5.1.3                         # Main framework
djangorestframework==3.15.1           # For building REST APIs
djangorestframework-simplejwt==5.3.1  # JWT-based authentication
django-cors-headers==4.6.0            # CORS middleware for handling cross-origin requests
django-filter==24.3                   # For filtering querysets in Django REST Framework
Pillow==11.0.0                        # For handling image files